MOVE: entity to another file
