+++ hermes/entities/applicant.py
@@ -17,12 +17,3 @@ class Applicant:
     image_path: str
 
 
-@dataclass
-class ApplicantStatus:
-    applicant_email: str
-    receipt_code: int
-    is_paid: bool
-    is_printed_application_arrived: bool
-    is_passed_first_apply: bool
-    is_final_submit: bool
-    exam_code: str
